third_party/WebKit/LayoutTests/fast/dom/private_script_unittest-expected.txt

Unittests for private scripts.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


This test needs window.internals.
PASS internals.return123() is 123
PASS internals.echoInteger(111) is 111
PASS internals.echoString("foo") is "foo"
PASS internals.addInteger(111, 222) is 333
PASS internals.addString("foo", "bar") is "foobar"
PASS internals.getIntegerFromPrototype() is 0
PASS internals.getIntegerFromPrototype() is 123
PASS internals.getIntegerFromDocument(document) is 0
PASS internals.getIntegerFromDocument(document) is 123
PASS internals.firstChild(node1) is node2
PASS internals.nextSibling(node2) is node3
PASS internals.nextSibling(node3) is node4
PASS internals.nextSibling(node4) is null
PASS internals.innerHTML(node5) is ""
PASS internals.innerHTML(node5) is "<div>foo</div>"
PASS internals.innerHTML(node6) is "foo"
PASS internals.innerHTML(node7) is ""
PASS internals.innerHTML(node7) is "clicked"
PASS internals.readonlyShortAttribute is 123
PASS internals.shortAttribute is -1
PASS internals.shortAttribute is 111
PASS internals.stringAttribute is "xxx"
PASS internals.stringAttribute is "foo"
PASS internals.nodeAttribute is null
PASS internals.nodeAttribute is node8
PASS successfullyParsed is true

TEST COMPLETE
